+++ backend/src/actor/actor.module.ts
@@ -5,9 +5,10 @@ import { ActorService } from './actor.service';
 import { ActorRepository } from './infrastructure/repository/actor.repository';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { MetadatasService } from 'src/metadatas/metadatas.service';
+import { ApplicationModule } from 'src/product/application.module';
 
 @Module({
-  imports: [PrismaModule],
+  imports: [PrismaModule, ApplicationModule],
   controllers: [ActorController],
   providers: [ActorService, ActorRepository, PrismaService, MetadatasService],
   exports: [ActorRepository],