+++ backend/src/common/utils/actionLog.utils.ts
@@ -1,0 +1,12 @@
+import { PrismaService } from 'src/prisma/prisma.service';
+
+const prisma: PrismaService = new PrismaService();
+
+export async function updateUserLastLogin(user: any) {
+  return await prisma.user.update({
+    where: { keycloakId: user.keycloakId },
+    data: {
+      lastLogin: new Date(),
+    },
+  });
+}