+++ backend/src/compliances/compliances.service.ts
@@ -3,10 +3,15 @@ import { PrismaService } from 'src/prisma/prisma.service';
 import { BaseService } from '../common/base.service';
 import { Compliance } from './entities/compliance.entity';
 import { MetadatasService } from 'src/metadatas/metadatas.service';
+import { ApplicationService } from 'src/product/application.service';
 
 @Injectable()
 export class CompliancesService extends BaseService<Compliance> {
-  constructor(prisma: PrismaService, metadatasService: MetadatasService) {
-    super(prisma.compliance, prisma, metadatasService);
+  constructor(
+    prisma: PrismaService,
+    metadatasService: MetadatasService,
+    applicationService: ApplicationService,
+  ) {
+    super(prisma.compliance, prisma, metadatasService, applicationService);
   }
 }