+++ backend/src/hosting/hosting.module.ts
@@ -5,9 +5,10 @@ import { PrismaModule } from '../prisma/prisma.module';
 import { HostingsController } from './hosting.controller';
 import { SitesController } from './site.controller';
 import { MetadatasModule } from 'src/metadatas/metadatas.module';
+import { ApplicationModule } from 'src/product/application.module';
 
 @Module({
-  imports: [PrismaModule, MetadatasModule],
+  imports: [PrismaModule, MetadatasModule, ApplicationModule],
   controllers: [HostingsController, SitesController],
   providers: [
     HostingService,