+++ backend/src/user/user.controller.ts
@@ -1,7 +1,16 @@
-import { Controller, Patch, Body, Param, Get } from '@nestjs/common';
+import {
+  Controller,
+  Patch,
+  Body,
+  Param,
+  Get,
+  Query,
+  Request,
+} from '@nestjs/common';
 import { UserService } from './user.service';
 import { ApiTags, ApiOperation, ApiParam, ApiResponse } from '@nestjs/swagger';
 import { UpdateUserDto } from './dto/update-user.dto';
+import { UserFilterDto } from './dto/filters.dto';
 import { Permissions } from '../common/decorators/permissions.decorator';
 
 @ApiTags('users')
@@ -9,13 +18,15 @@ import { Permissions } from '../common/decorators/permissions.decorator';
 export class UserController {
   constructor(private readonly userService: UserService) {}
 
-  @Get(':id')
-  @ApiOperation({ summary: 'Récupérer un utilisateur par ID' })
-  @ApiParam({ name: 'id', description: "ID Keycloak de l'utilisateur" })
-  @ApiResponse({ status: 200, description: 'Utilisateur trouvé' })
+  @Get('me')
+  @ApiOperation({ summary: 'Récupérer ses propres informations utilisateur' })
+  @ApiResponse({
+    status: 200,
+    description: 'Informations utilisateur trouvées',
+  })
   @ApiResponse({ status: 404, description: 'Utilisateur non trouvé' })
-  async findOne(@Param('id') id: string) {
-    return this.userService.findUserByKeycloakId(id);
+  async findMe(@Request() req: any) {
+    return req.user;
   }
 
   @Patch(':id')
@@ -44,14 +55,14 @@ export class UserController {
   @ApiOperation({
     summary: 'Lister tous les utilisateurs',
     description:
-      'Récupère la liste de tous les utilisateurs avec leurs permissions. Accès limité aux administrateurs.',
+      'Récupère la liste de tous les utilisateurs avec leurs permissions. Supporte la recherche par email et ID Keycloak. Accès limité aux administrateurs.',
   })
   @ApiResponse({ status: 200, description: 'Liste des utilisateurs' })
   @ApiResponse({
     status: 403,
     description: 'Accès refusé - Privilège admin requis',
   })
-  async findAll() {
-    return this.userService.findAll();
+  async findAll(@Query() filters: UserFilterDto) {
+    return this.userService.findAll(filters);
   }
 }